hub: split input staging out of relay.rs
No behavior change.
